Day 06 - Custom Customs
I got the answer right about two hours ago. The rest of time I spent trying to understand why the number expected for the result was diffent from the one I had got... which didn't make any sense since I was sure about what I had written... turns out the last group of the input had an extra empty line. And an intersection of the letters of a string and an empty string, is always an empty string. Because of this crap the number I got was 6, bloody 6, lower than the expected number. Just venting for posterity. My repos are basically > /dev/null